Move seeding tests functionality into test module
